+++ app/models/folio/attribute_type.rb
@@ -14,15 +14,21 @@ class Folio::AttributeType < Folio::ApplicationRecord
                               dependent: :destroy,
                               foreign_key: :folio_attribute_type_id
 
-  validates :title,
-            presence: true,
-            uniqueness: { scope: %i[type site_id] }
-
   validates :type,
             presence: true
 
   validates :data_type,
             inclusion: { in: DATA_TYPES }
+
+  if Rails.application.config.folio_using_traco
+    translates :title
+
+    validate :validate_title_for_site_locales
+  else
+    validates :title,
+              presence: true,
+              uniqueness: { scope: %i[type site_id] }
+  end
 
   def data_type_with_default
     data_type || DATA_TYPES.first
@@ -37,6 +43,25 @@ class Folio::AttributeType < Folio::ApplicationRecord
       [human_attribute_name("data_type/#{data_type}"), data_type]
     end
   end
+
+  private
+    def validate_title_for_site_locales
+      if site.blank?
+        errors.add(:site, :blank)
+      else
+        site.locales.each do |locale|
+          title_attr = "title_#{locale}"
+
+          if send(title_attr).blank?
+            errors.add(title_attr, :blank)
+          elsif self.class.where.not(id:)
+                          .where("title_#{locale}" => send(title_attr), type:, site_id:)
+                          .exists?
+            errors.add(title_attr, :taken)
+          end
+        end
+      end
+    end
 end
 
 # == Schema Information